Remove debug log

+++ controllers/commentsController.js
@@ -81,11 +81,9 @@ exports.removeCommentAdmin = async (req, res) => {
       id: user.id,
     }
   })
-
-  console.log(userDb.role);
-  // if (userDb.role !== "ADMIN") {
-  //   return res.sendStatus(403);
-  // }
+  if (userDb.role !== "ADMIN") {
+    return res.sendStatus(403);
+  }
   await prisma.comment.delete({
     where: {
       id: commentId
